backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
spacy
fuzzywuzzy[speedup]
PyMuPDF